+++ ChangeLog.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `e3nn.SphericalSignal.sample` to sample a point on the sphere
+
+### Fixed
+- `e3nn.SphericalSignal` arithmetic operations
 
 ## [0.16.0] - 2022-02-01
 ### Added